Create SolveASudoku10112024.py
A first example that you can hand-edit for your own puzzle. This should work on any python system, but was tested for use at https://www.online-python.com/ - you can just copy and paste the code into their editor and run it. Their run system has trouble with having a live prompt for a long time, so you will have to interact with it quickly. I recommend just using option 2 as soon as the program starts in order to solve the puzzle and print all of the output lines. You should modify option 2 to repeat as many times as necessary to solve the problem or have a full cycle (1-9) where no new digits are added. You can improve on this program by (A) making the algorithm in option 2 automatically figure out if it solves the puzzle or if the puzzle gets stuck, (B) adding new solving methods, and (C) improving the interaction with the user - i.e. tell the user how many new digits were found for each test. Note that it is recommended to move this to your own python environment so that you can more easily control your development. However, the website (online-python.com) does a great job of context-sensitive help so that the code is easier to write than a plain text editor.